requirements.txt
djangorestframework>=2.4.3
geopy>=1.11.0
requests>=2.13.0
